Move helper functions back into replicaset util
