+++ controller/sync.go
@@ -204,14 +204,15 @@ func (c *Controller) getNewReplicaSet(rollout *v1alpha1.Rollout, rsList, oldRSs 
 	return createdRS, err
 }
 
-// sync is responsible for reconciling rollouts on scaling events.
-func (c *Controller) sync(r *v1alpha1.Rollout, rsList []*appsv1.ReplicaSet) error {
+// syncScalingEvent is responsible for reconciling rollouts on scaling events.
+func (c *Controller) syncScalingEvent(r *v1alpha1.Rollout, rsList []*appsv1.ReplicaSet) error {
 	logCtx := logutil.WithRollout(r)
 	logCtx.Info("Reconciling scaling event")
 	newRS, oldRSs, err := c.getAllReplicaSetsAndSyncRevision(r, rsList, false)
 	if err != nil {
 		return err
 	}
+	// NOTE: it is possible for newRS to be nil (e.g. template and replicas changed at same time)
 	if r.Spec.Strategy.BlueGreenStrategy != nil {
 		previewSvc, activeSvc, err := c.getPreviewAndActiveServices(r)
 		if err != nil {
@@ -302,10 +303,11 @@ func (c *Controller) calculateBaseStatus(allRSs []*appsv1.ReplicaSet, newRS *app
 
 	var currentPodHash string
 	if newRS == nil {
-		// newRS potentially might be nil when blue-green rollout receives a scaling event in
-		// Controller::sync(). This is probably dead code and ideally should be removed eventually.
-		logutil.WithRollout(rollout).Warn("Unexpected nil for newRS")
+		// newRS potentially might be nil when called by Controller::syncScalingEvent(). For this
+		// to happen, the user would have had to simultaneously change to change the number of
+		// replicas, and the pod template spec at the same time.
 		currentPodHash = controller.ComputeHash(&rollout.Spec.Template, rollout.Status.CollisionCount)
+		logutil.WithRollout(rollout).Warnf("Assuming %s for new replicaset pod hash", currentPodHash)
 	} else {
 		currentPodHash = newRS.Labels[v1alpha1.DefaultRolloutUniqueLabelKey]
 	}
