+++ utils/replicaset/replicaset.go
@@ -14,9 +14,11 @@ import (
 	intstrutil "k8s.io/apimachinery/pkg/util/intstr"
 	corev1defaults "k8s.io/kubernetes/pkg/apis/core/v1"
 	"k8s.io/kubernetes/pkg/controller"
+	"k8s.io/utils/pointer"
 
 	v1alpha1 "github.com/argoproj/argo-rollouts/pkg/apis/rollouts/v1alpha1"
 	"github.com/argoproj/argo-rollouts/utils/annotations"
+	"github.com/argoproj/argo-rollouts/utils/conditions"
 	"github.com/argoproj/argo-rollouts/utils/defaults"
 	logutil "github.com/argoproj/argo-rollouts/utils/log"
 )
@@ -36,7 +38,7 @@ func FindNewReplicaSet(rollout *v1alpha1.Rollout, rsList []*appsv1.ReplicaSet) *
 	// This covers the corner case in which the reason we did not find the replicaset, was because
 	// of a change in the controller.ComputeHash function (e.g. due to an update of k8s libraries).
 	// When this (rare) situation arises, we do not want to return nil, since nil is considered a
-	// PodTempalte change, which in turn would triggers an unexpected redeploy of the replicaset.
+	// PodTemplate change, which in turn would triggers an unexpected redeploy of the replicaset.
 	for _, rs := range rsList {
 		if PodTemplateEqualIgnoreHash(&rs.Spec.Template, &rollout.Spec.Template) {
 			logCtx := logutil.WithRollout(rollout)
@@ -241,6 +243,60 @@ func MaxSurge(rollout *v1alpha1.Rollout) int32 {
 	// Error caught by validation
 	maxSurge, _, _ := resolveFenceposts(defaults.GetMaxSurgeOrDefault(rollout), defaults.GetMaxUnavailableOrDefault(rollout), rolloutReplicas)
 	return maxSurge
+}
+
+// checkStepHashChange indicates if the rollout's step for the strategy have changed. This causes the rollout to reset the
+// currentStepIndex to zero. If there is no previous pod spec to compare to the function defaults to false
+func checkStepHashChange(rollout *v1alpha1.Rollout) bool {
+	if rollout.Status.CurrentStepHash == "" {
+		return false
+	}
+	// TODO: conditions.ComputeStepHash is not stable and will change
+	stepsHash := conditions.ComputeStepHash(rollout)
+	if rollout.Status.CurrentStepHash != conditions.ComputeStepHash(rollout) {
+		logCtx := logutil.WithRollout(rollout)
+		logCtx.Infof("Canary steps change detected (new: %s, old: %s)", stepsHash, rollout.Status.CurrentStepHash)
+		return true
+	}
+	return false
+}
+
+// checkPodSpecChange indicates if the rollout spec has changed indicating that the rollout needs to reset the
+// currentStepIndex to zero. If there is no previous pod spec to compare to the function defaults to false
+func checkPodSpecChange(rollout *v1alpha1.Rollout, newRS *appsv1.ReplicaSet) bool {
+	if rollout.Status.CurrentPodHash == "" {
+		return false
+	}
+	podHash := GetPodTemplateHash(newRS)
+	if rollout.Status.CurrentPodHash != podHash {
+		logCtx := logutil.WithRollout(rollout)
+		logCtx.Infof("Pod template change detected (new: %s, old: %s)", podHash, rollout.Status.CurrentPodHash)
+		return true
+	}
+	return false
+}
+
+// PodTemplateOrStepsChanged detects if there is a change in either the pod template, or canary steps
+func PodTemplateOrStepsChanged(rollout *v1alpha1.Rollout, rsList []*appsv1.ReplicaSet) bool {
+	newRS := FindNewReplicaSet(rollout, rsList)
+	if newRS == nil {
+		return false
+	}
+	if checkPodSpecChange(rollout, newRS) {
+		return true
+	}
+	if checkStepHashChange(rollout) {
+		return true
+	}
+	return false
+}
+
+// ResetCurrentStepIndex resets the index back to zero unless there are no steps
+func ResetCurrentStepIndex(rollout *v1alpha1.Rollout) *int32 {
+	if len(rollout.Spec.Strategy.CanaryStrategy.Steps) > 0 {
+		return pointer.Int32Ptr(0)
+	}
+	return nil
 }
 
 // PodTemplateEqualIgnoreHash returns true if two given podTemplateSpec are equal, ignoring the diff in value of Labels[pod-template-hash]
